+++ CHANGELOG.rst
@@ -4,11 +4,17 @@ Changelog
 Development
 ===========
 
+...
+
+0.7.0 (16.09.2020)
+==================
+
 - Add test for Jupyter notebook
 - Add function to discover available climate observations
   (time resolution, parameter, period type)
 - Make the CLI work again and add software tests to prevent future havocs
 - Use Sphinx Material theme for documentation
+- Fix typo in enumeration for TimeResolution.MINUTES_10
 
 0.6.0 (07.09.2020)
 ==================